cppcheck warning reduction/improvements
- Change sim_emax to size_t, since it's used as an array offset. This
  change ripples through the simulator, unfortunately.
- Fix shadowed variables
- Reduce variable scopes where suggested.
- Fix "x & y ? r1 : r2" precedence issues.
- Reformat code to make it readable.

- Fixed WITH_NETWORK=Off issue on Win32: Really should link with the
  WinXX socket libraries because sim_ether.c wants it.

+++ 3B2/3b2_sys.c
@@ -47,7 +47,7 @@ REG *sim_PC = &cpu_reg[0];
 
 /* All opcodes are 1 or 2 bytes. Operands may be up to 6 bytes, and
    there may be up to 3 operands, for a maximum of 20 bytes */
-int32 sim_emax = 20;
+size_t sim_emax = 20;
 
 extern instr *cpu_instr;
