+++ CDC1700/cdc1700_sys.c
@@ -65,7 +65,7 @@ char sim_name[] = "CDC1700";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 2;
+size_t sim_emax = 2;
 
 DEVICE *sim_devices[] = {
   &cpu_dev,
