+++ I7000/i7010_sys.c
@@ -39,7 +39,7 @@ char                sim_name[] = "IBM 7010";
 
 REG                *sim_PC = &cpu_reg[0];
 
-int32               sim_emax = 50;
+size_t              sim_emax = 50;
 
 DEVICE             *sim_devices[] = {
     &cpu_dev,
