+++ PDP1/pdp1_sys.c
@@ -82,7 +82,7 @@ char sim_name[] = "PDP-1";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 1;
+size_t sim_emax = 1;
 
 DEVICE *sim_devices[] = {
     &cpu_dev,
