+++ TX-0/tx0_sys.c
@@ -73,7 +73,7 @@ char sim_name[] = "TX-0";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 1;
+size_t sim_emax = 1;
 
 DEVICE *sim_devices[] = {
     &cpu_dev,
